Created a new Android Library module "azure-storage-common" and moved the current response interceptors there. Renamed the azure-core interceptors package to interceptor.
