+++ sdk/settings.gradle
@@ -1,1 +1,1 @@
-include ':core:azure-core'
+include ':core:azure-core', ':storage:azure-storage-common'
