+++ sdk/storage/azure-storage-common/src/main/java/com/azure/android/storage/common/interceptor/ResponseValidationInterceptor.java
@@ -1,4 +1,4 @@
-package com.azure.android.core.http.interceptors;
+package com.azure.android.storage.common.interceptor;
 
 import androidx.annotation.NonNull;
