+++ sdk/storage/azure-storage-common/src/main/java/com/azure/android/storage/common/interceptor/ScrubEtagInterceptor.java
@@ -1,4 +1,4 @@
-package com.azure.android.core.http.interceptors;
+package com.azure.android.storage.common.interceptor;
 
 import androidx.annotation.NonNull;
 
@@ -8,7 +8,7 @@ import okhttp3.Interceptor;
 import okhttp3.Response;
 
 /**
- * Wraps any potential error responses from the service and applies post processing of the response's eTag header to
+ * Wraps any potential error responses from the service and applies post processing of the response's ETag header to
  * standardize the value.
  */
 public class ScrubEtagInterceptor implements Interceptor {
